Ensure that SegmentedQueue retains static blocks
New dynamic blocks can be inserted anywhere in the block list, but they
were only removed from the end. This meant that we could be left with
different blocks than originally allocated. If dynamic blocks were
allocated from DRAM due to transient SRAM heap pressure, we could be
left with a DRAM-based block as the primary block, resulting in a crash.

Prevent this condition by remembering which blocks are static, and never
deallocating them. Ideally these would be kept in a fixed-size array,
but since the static block count is not a template parameter, put it in
a DynamicVector instead. Offset some of the increase memory requirement
by consolidating the max & static block count constants into their
vectors.

Bug: 405295187
Test: ./run_tests.sh
Change-Id: I64052e4b78f82a288e9c336450618c2795a67847