[CHPP] Log error during longer than expected wait time during sync request
If there is a bug in the code, it's possible that we may be waiting longer than expected for a sync response to arrive/time out. Log an error in this case for easier debugging, and break the for loop to avoid blocking the client thread unnecessarily.

Bug: 416780648
Flag: EXEMPT minor bug fix
Test: Smoke test on device
Change-Id: I8042a788caa302e4c6509498a5c957e7c1e893e8