Update pw::Allocator implementations
Move the to pigweed subdirectory for better grouping, and rename to reflect
that it's a pw::Allocator, not a container allocator.

Also add a pw::Allocator concrete implementation that maps to nanoapp
allocation functions, for situations where nanoapps want to use Pigweed
containers or other components that take a pw::Allocator argument, and
they'd prefer to map this into the CHRE heap allocation functions rather
than some platform-specific or local allocator.

Bug: 422577795
Test: atest chre_unit_tests + chre_simulation_tests
Change-Id: I5e16c44d536a43590195cd69bdcd95fa733865fe